FIXES-30

+++ database/migrations/2023_08_08_051827_create_performa_invoices_table.php
@@ -24,7 +24,7 @@ class CreatePerformaInvoicesTable extends Migration
             $table->string('mode', 100);
             $table->string('job_number', 100);
             $table->string('promo_code', 100);
-            $table->float('tax', 10, 0);
+            $table->decimal('tax', 10, 2);
             $table->text('tax_selected')->nullable();
             $table->integer('total');
             $table->string('invoice_type', 100);
